+++ src/index.ts
@@ -1,3 +1,3 @@
+export * from './misc/providers';
 export * from './misc/serverRenderFactory';
 export * from './misc/ssrProgressIndicator.factory';
-export * from './modules/serverProviders.module';